Add character generation based on probabilities
In order to generate words that sounds correct for a given referential
(it may be french words, english cities, stars names), we must have the
ability to choose a character based on the number of times they appear
in that referential.

For instance, if we want to generate some french words, because the
letter 'e' appears way more than the letter 'z', we want 'e' generated
more times.